[김주영 상영관 삭제 기능 상영스케쥴과 연동 처리]

+++ CinepickProject/src/main/java/com/itwillbs/cinepick/service/AdminService.java
@@ -345,6 +345,12 @@ public class AdminService {
 	//상영관 삭제
 	public int deleteScreen(int screenIdx) {
 		System.out.println("AdminService - deleteScreen()");
+		
+		int existCount = mapper.checkExistScreen(screenIdx);
+		
+		if (existCount > 0) {
+			return -1;
+		}
 		return mapper.deleteScreen(screenIdx);
 	}
